Move RestXxx functions into src/api/rest (2nd commit)
