+++ src/api/rest/server/common/namespace.go
